requirements.txt
# Básicos
streamlit
pandas
numpy<2.0.0

# Visualización
plotly
openpyxl

# Base de datos
pymongo

# Google APIs
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
google-api-python-client

# Procesamiento de datos
xlsxwriter
streamlit-aggrid

# Utilidades
python-dateutil
tqdm

prophet
scikit-learn

# Para análisis predictivo
prophet==1.1.4
statsmodels==0.14.0
scikit-learn>=1.0.2

pytz

pymongo==4.6.1
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.115.0